// frontend-mobile/src/lib/types.ts
export interface Request {
  id: string;
  title: string;
  content: string;
  category: 'Counseling' | 'Prayer' | 'Complaint' | 'Suggestion';
  status: 'Received' | 'Assigned' | 'In Progress' | 'Resolved';
  submittedBy: string;
  assignedTo: string;
  addedBy: string;
  date: string;
}